Extract test generation logic in dedicated module
