+++ tests/common/mod.rs
@@ -1,0 +1,167 @@
+use gpex::Log;
+use itertools::{Itertools, MultiProduct};
+use regex::Regex;
+use serde::Deserialize;
+use std::collections::HashMap;
+use std::ffi::OsStr;
+use std::fs::File;
+use std::path::{Path, PathBuf};
+use std::vec::IntoIter;
+use std::{env, fs, io};
+
+const CASE_KEY_PLACEHOLDER: &str = "$$";
+
+#[derive(Debug)]
+#[expect(dead_code)] // variant data is useful for debugging tests
+pub(crate) enum Error {
+    Io(io::Error),
+    Regex(regex::Error),
+    Yaml(serde_yml::Error),
+    Gpex(Vec<Log>),
+}
+
+#[derive(Debug, Deserialize)]
+struct Cases {
+    dimensions: Vec<Dimension>,
+    #[serde(default)]
+    exclusions: Vec<HashMap<String, Vec<String>>>,
+}
+
+#[derive(Debug, Deserialize)]
+struct Dimension {
+    id: String,
+    cases: HashMap<String, HashMap<String, String>>,
+}
+
+#[derive(Debug, Clone)]
+struct DimensionCase {
+    dimension: String,
+    name: String,
+    key_values: HashMap<String, String>,
+}
+
+pub(crate) fn generate_cases(path: &Path) -> Result<PathBuf, Error> {
+    let cases_file = path.join("cases.yaml");
+    if cases_file.exists() {
+        let test_dir = env::temp_dir().join(path);
+        _ = fs::remove_dir_all(&test_dir);
+        let cases_file = File::open(cases_file).map_err(Error::Io)?;
+        let cases: Cases = serde_yml::from_reader(cases_file).map_err(Error::Yaml)?;
+        let case_combinations = case_combinations(&cases)
+            .filter(|combination| !is_case_combination_excluded(&cases, combination))
+            .collect::<Vec<_>>();
+        generate_dir(path, &case_combinations)?;
+        Ok(test_dir)
+    } else {
+        Ok(path.to_path_buf())
+    }
+}
+
+fn case_combinations(cases: &Cases) -> MultiProduct<IntoIter<DimensionCase>> {
+    cases
+        .dimensions
+        .iter()
+        .map(|dimension| {
+            dimension
+                .cases
+                .iter()
+                .map(|(key, value)| DimensionCase {
+                    dimension: dimension.id.clone(),
+                    name: key.clone(),
+                    key_values: value.clone(),
+                })
+                .collect::<Vec<_>>()
+        })
+        .multi_cartesian_product()
+}
+
+fn is_case_combination_excluded(cases: &Cases, combination: &[DimensionCase]) -> bool {
+    for exclusion in &cases.exclusions {
+        if exclusion.iter().all(|(dimension, cases)| {
+            cases.contains(find_case_from_dimension(combination, dimension))
+        }) {
+            return true;
+        }
+    }
+    false
+}
+
+fn find_case_from_dimension<'a>(
+    combination: &'a [DimensionCase],
+    dimension: &String,
+) -> &'a String {
+    &combination
+        .iter()
+        .find(|case| &case.dimension == dimension)
+        .unwrap_or_else(|| panic!("cases.yaml: '{dimension}' dimension not found"))
+        .name
+}
+
+fn generate_dir(current_path: &Path, cases: &[Vec<DimensionCase>]) -> Result<(), Error> {
+    for entry in current_path.read_dir().map_err(Error::Io)? {
+        let entry = entry.map_err(Error::Io)?;
+        let path = entry.path();
+        let file_type = entry.file_type().map_err(Error::Io)?;
+        if file_type.is_dir() {
+            generate_dir(&path, cases)?;
+        } else if path.extension() == Some(OsStr::new("gpex")) {
+            generate_file(&path, cases)?;
+        }
+    }
+    Ok(())
+}
+
+fn generate_file(file_path: &Path, cases: &[Vec<DimensionCase>]) -> Result<(), Error> {
+    if !is_path_containing(file_path, CASE_KEY_PLACEHOLDER) {
+        let output_path = env::temp_dir().join(file_path);
+        // Helps to detect missing case placeholder in case the file includes key placeholders.
+        fs::create_dir_all(path_parent(&output_path)).map_err(Error::Io)?;
+        fs::copy(file_path, &output_path).map_err(Error::Io)?;
+        return Ok(());
+    }
+    let placeholder_regex = Regex::new(r"\{\{.*}}").map_err(Error::Regex)?;
+    for case in cases {
+        let case_name = case.iter().map(|dimension| &dimension.name).join("__");
+        let mut content = fs::read_to_string(file_path)
+            .map_err(Error::Io)?
+            .replace(CASE_KEY_PLACEHOLDER, &case_name);
+        for dimension in case {
+            for (key, value) in &dimension.key_values {
+                content =
+                    content.replace(&format!("{{{{{}.{}}}}}", dimension.dimension, key), value);
+            }
+        }
+        let output_path =
+            env::temp_dir().join(replace_in_path(file_path, CASE_KEY_PLACEHOLDER, &case_name));
+        content = placeholder_regex.replace_all(&content, "").to_string();
+        fs::create_dir_all(path_parent(&output_path)).map_err(Error::Io)?;
+        fs::write(&output_path, &content).map_err(Error::Io)?;
+    }
+    Ok(())
+}
+
+fn path_parent(path: &Path) -> &Path {
+    path.parent()
+        .unwrap_or_else(|| unreachable!("parent should be at least tmp folder"))
+}
+
+fn is_path_containing(file_path: &Path, pattern: &str) -> bool {
+    file_path.iter().any(|component| {
+        component
+            .to_str()
+            .unwrap_or_else(|| unreachable!("invalid test file path"))
+            .contains(pattern)
+    })
+}
+
+fn replace_in_path(file_path: &Path, pattern: &str, replacement: &str) -> PathBuf {
+    file_path
+        .iter()
+        .map(|component| {
+            component
+                .to_str()
+                .unwrap_or_else(|| unreachable!("invalid test file path"))
+                .replace(pattern, replacement)
+        })
+        .collect::<PathBuf>()
+}
